+++ packages/dash-html-components/tests/test_dash_import.py
@@ -5,18 +5,18 @@ import unittest
 class TestDashImport(unittest.TestCase):
     def setUp(self):
         with open('dash.py', 'w') as f:
-            pass        
-        
+            pass
+
     def tearDown(self):
         try:
             os.remove('dash.py')
             os.remove('dash.pyc')
         except OSError:
             pass
-        
+
     def test_dash_import(self):
         """Test that program exits if the wrong dash module was imported"""
-        
+
         with self.assertRaises(SystemExit) as cm:
             import dash_html_components